MueLu: removed "repartition: keep proc 0" option
and corresponding code in RepartitionFactory.

Note: parallel interface tests now miss a chunk of output, as only
processor 0 can write into a file. This could be fixed:
a) write a script that calls the executable redirecting the output into
   a file. This would rely on MPI combining std::cout output, and using
   system utilities to redirect that. Pretty easy to implement.
b) upgrading FancyOStream to work with parallel file IO
c) make each processor write into a local buffer, and combine those
   buffers at the end of the run. Problematic, as combining would
   require timestamps, or something similar.

I think a) is the way to go, and should be easy enough to implement.