Xpetra: Fixed build errors if KokkosClassic::SerialNode is OFF.
It is now possible to build Xpetra if Tpetra is ON but
KokkosClassic::SerialNode is OFF.

I also fixed the Cloner unit tests so that they try their best to
exercise clone() for two different Node types.  They do this by
attempting to choose a second Node type that differs from the default
Node type.  If only one Node type is enabled, it falls back to calling
clone() with the same Node type as the default.  The logic for picking
the second Node type uses a long nest of C preprocessor conditionals.
It's ugly, but appears to work.