Add specialzation for unsigned __int64 to address Intel Windows build issue (Trilinos #6239)
In earlier commit by Jordan, a specialization for size_t was added that
conflicited with other specializations on other platforms.  That lead the
commit being reverted.

This new commit changes the type to a fundamental type 'unsigned __int64' but
only if that type exists (as already being determined by the configure-time
test that defines HAVE_TEUCHOS___INT64.

This should hopefully resolve the Windows build issue for SCALE case 7337.