Kokkos: updated nvcc_wrapper
The nvcc_wrapper now accepts --std=c++11 to enable C++11 support.
The nvcc_wrapper now has an option --replace-pragma-ident which if
enabled will replace on the fly #pragma ident with #ident. Both
appear to be non standard functionality, and while nvcc can accept
it can't take
while gcc can.
What nvcc can do is
as well as

This issue occured when compiling OpenMPI which in its configure
script tests for
but in the source code uses