KokkosAlgorithms: Fixed Kokkos::rand unit test.
Kokkos::rand<Generator, Scalar>::max() is supposed to define an
exclusive upper bound on the range of random numbers that
Kokkos::rand<Generator, Scalar>::draw() can generate.  However, for
the Scalar = float specialization, the current implementation violates
the exclusiveness of this upper bound, due to rounding error.  (That
is, max() returns 1.0f, but draw() may also return 1.0f.  I've seen it
do so.)  It's entirely likely that third-party library implementations
might violate this as well.  This breaks the current test for
Kokkos::rand, because the histogram dimensions assume that the upper
bound is exclusive.  The error only shows up deterministically if
Kokkos' run-time bounds checking is turned on.

I fixed this by leaving an extra entry at the end of each dimension of
the histograms.  I also took the opportunity to add a Kokkos::Serial
specialization of the Kokkos::rand test.