TeuchosNumerics: Fix for Bug 5968.
Teuchos::SerialDense{Matrix,Vector} and related classes no longer have
their Object label set by default.  Before, every time an instance of
one of these classes was created or copied, the Object's label would
be set.  This invokes strcpy(), which can be expensive for a large
number of very small matrices and vectors.

My guess is that Teuchos::Object was meant to serve as a partial
replacement for RTTI (Run-Time Type Information), on platforms that
did not support RTTI.  The Object label was always being set to the
class' name.  We can use RTII (e.g., std::typeinfo) to extract that
information at run time, so there's no need for an Object base class
to remember the class' name as a string.  I'm leaving Object in place
for backwards compatibility, but it really should go away at some
point.