Sacado:  Fix Promote<> spec for unary expressions.
My default Promote specializations using is_convertible<> didn't work
for unary expressions because the AD type is convertible to the
expression in simple cases.  I added a Promote specialization when both
A and B are convertible to each other to handle this case.  Also added
unit tests for this case.