Tpetra: Got ReplaceLocalValues test to pass, through some trickery.
With the Kokkos refactor version of Tpetra enabled, with Tpetra's
default Node type set to Kokkos::Compat::KokkosSerialWrapperNode, and
with the OpenMP device apparently enabled (though not explicitly in
the build configuration!), the CrsMatrix_ReplaceLocalValues test was
failing with an exception.  The exception claims that some code is
calling parallel_for on the Kokkos::OpenMP device, without that device
having been initialized.  This is a bit weird because the default Node
type in this case should _only_ use the Kokkos::Serial device; it
should not use the OpenMP device at all.  Furthermore, the Map
constructor should be creating the Node instance, and that should
trigger device initialization.

I was able to make the test pass by letting it build and run for _all_
enabled Node types, not just for the default Node type.  This must
have something to do with the order in which Node types are
initialized.  I added some checks to the test for whether that device
and others are initialized, but the checks didn't help me catch the
issue.  The test is better off exercising all Node types anyway, so I
will leave it in place.