[Nox] Remove redundant printUpdate().
On iteration 0, if already converged, output was being written twice. That was
because printUpdate() was called twice. Analysis of the code showed that if the
first one is issued, then the second must be, as well: hence the redundancy. I
removed the first iteration-0 specific one to remove the redundancy.