Sacado:  Fix SFINAE logic for +=,-=,*=,/- with expression r.h.s.
I had a cut-and-paste error for the SFNIAE logic for enabling the
overloads of +=,-=,*=,/= which was preventing these overlaods from ever
being used.  Not only is this ineffecient since it results in the
creation of a temporary, but also incorrect for ViewFad since creating
temporary ViewFad's isn't allowed.  I used the correct SFINAE macro to
fix this and added a unit test as well.  I also removed all of the
normal constructors for ViewFad to prevent temporaries from ever being
created.  This required a small update to the Promote specializations as
well as the conversion unit test.