Kokkos: Moved FENL example from KokkosExample to KokkosLinAlg.
The FENL example depends on KokkosCore, KokkosContainers, and
KokkosLinAlg.  No other examples in the KokkosExample subpackage
depend on KokkosLinAlg.  Since we want to move KokkosLinAlg out of
Kokkos and into Tpetra, as a first step, I have moved the FENL example
out of KokkosExample and into KokkosLinAlg.

I also changed KokkosExample's dependencies so that it no longer
depends on KokkosLinAlg.