Tpetra::CrsGraph: Got rid of numAllocPerRow_ legacy array.
This commit only affects the Kokkos refactor version of Tpetra.

numAllocPerRow_ stores the maximum number of entries in each locally
owned row.  In the Kokkos refactor version of CrsGraph, it is a legacy
backwards compatibility data structure which views the Kokkos array
k_numAllocPerRow_.  This commit removes it from the Kokkos refactor
version of CrsGraph.