Ifpack2: Added ETI for RowMatrix version of OneLevelFactory.
Previously, Ifpack2::Details::OneLevelFactory only included explicit
template instantiation (ETI) for when the template parameter is a
Tpetra::CrsMatrix specialization.  This caused a link error when
building MueLu, with the Kokkos refactor version of Tpetra enabled, when
KokkosThreadsWrapperNode is enabled but NOT the default Node.  To fix
this temporarily, I added ETI for when the template parameter is a
Tpetra::RowMatrix specialization.  I really think the RowMatrix
specialization should be the only one ever used, but we can deal with
that later.