Ifpack2::Chebyshev: Made power method robust to unlucky initial guess.
With the Kokkos refactor version of Tpetra enabled and
Kokkos::Compat::KokkosThreadsWrapperNode as the default Node type, the
Ifpack2_Cheby_belos test was failing.  This was because Chebyshev
iteration took 37 iterations to converge on the given test problem,
which is greater than the maximum allowed of 32.  The "classic" version
of Tpetra (Node = KokkosClassic::SerialNode) was not failing, and indeed
took only 19 Chebyshev iterations to converge.

I found out that the difference was entirely due to the pseudorandom
initial vector in Chebyshev's power method iteration.  By default,
Chebyshev uses a fixed number of power method iterations to guess the
maximum eigenvalue of the scaled matrix.  When I set the initial guess
vector of the power method to all ones, Chebyshev converged in 28
iterations for both versions of Tpetra.  The Kokkos refactor version of
Tpetra uses a different pseudorandom number generator for
Tpetra::MultiVector::randomize(), so one should expect some difference
between the two Tpetra versions.  In this case, the Kokkos refactor
version happened to catch an unlucky case.  I specifically wrote
"unlucky" rather than "bad," because Kokkos refactor Tpetra has a better
pseudorandom number generator than classic Tpetra.

The power method iteration in Ifpack2 forms the maximum eigenvalue
estimate as the ratio of a dot product and a norm squared.  Negative
values are possible for the first, and I observed them in intermediate
iterations with both versions of Tpetra.  With the Kokkos refactor
version of Tpetra, though, the default number of iterations was not
enough to make the max eigenvalue estimate positive.  Chebyshev
iteration is sensitive to the eigenvalue estimates, so this is likely
the reason why it was not converging fast enough.

I would very much like the Ifpack2 unit test to be independent of
Tpetra's choice of pseudorandom numbers.  The KokkosAlgorithms
subpackage already tests the pseudorandom number generation that the
Kokkos refactor version of Tpetra uses, so we don't need to make Ifpack2
test it.  I also don't want to relax the maximum number of Chebyshev
iterations the test allows without a good reason.  I could just add an
option to Ifpack2::Chebyshev to control the initial guess for the power
method.  For example, the option could force a specific nonrandom
initial guess (e.g., a vector of all ones), or it could even let the
user provide an initial guess vector, just as Chebyshev currently lets
users provide an initial guess for Chebyshev iteration itself.  However,
I don't like to add rarely used options that would not get enough
testing.  I would also like to make Ifpack2 more robust to "unlucky"
random number generation like this.

Ifpack2's Chebyshev implementation currently only works for real,
symmetric positive definite matrices.  We can exploit this to make the
power method more robust.  In particular, if the current max eigenvalue
estimate is negative after the maximum iteration count, it might be
worthwhile to draw some more random numbers and try again.  This commit
implements that solution: if the first try doesn't work, the power
method draws another initial guess vector and tries again.  It only does
this once (that is, it tries a maximum of two initial guess vectors).
That keeps the run time bounded while making the power method more
robust to an unlucky draw of pseudorandom numbers.