TpetraClassic: Stop promoting KokkosClassic namespace into Kokkos.
A while back, we started promoting all the contents of the KokkosClassic
namespace into the Kokkos namespace.  That was a temporary backwards
compatibility measure.  This commit stops the promotion, so that
"Kokkos::" no longer works as an "alias" for "KokkosClassic::".  This
helps prepare for the public release of the Kokkos refactor version of
Tpetra.

Almost all Trilinos packages already use "KokkosClassic::" correctly.  I
only made one change in PyTrilinos, where Kokkos::DefaultNode was being
used instead of KokkosClassic::DefaultNode.  Please note that it is
better to rely on the default value of the "Node" template parameter in
Tpetra classes.  You can access the default Node type using an expresion
like Tpetra::Map<>::node_type.