Tpetra: Deprecated HybridPlatform.
HybridPlatform was meant to let users run their code (templated on Node
type) with any enabled Node type.  "Hybrid" meant more than just "MPI+X
for some fixed X"; it meant "MPI+X for all enabled X, where X may differ
between MPI processes."  This was a strategy for making use of both GPUs
and multicore CPUs on the same node.  It was also useful for performance
comparisons.

HybridPlatform is fragile to maintain.  This is because it requires a
list of all existing Node types at compile time, in order to map from
run-time parameters (in a Teuchos::ParameterList) to the Node type.
Kokkos has a better way of dealing with heterogeneous execution
resources on the same node, that Tpetra is currently implementing and
extending.  HybridPlatform also saw very little use.  Many users found
its name confusing; they would go to that class, looking for MPI+X
support for some fixed X, but found something more complicated that they
didn't understand.

For all these reasons, I am deprecating this class, and removing its
tests.

Some Tpetra RTI examples use HybridPlatform, but this is orthogonal to
RTI's purpose.  Furthermore, RTI does not currently work with the Kokkos
refactor version of Tpetra.  This should not be hard to fix, but it
means that RTI is disabled by default, so we do not need to worry about
deprecation warnings.