Sacado:  Fix ViewFadType to not put const in the base Fad type.
The last template parameter of ViewFad contains the original Fad type
used in the view, and ViewFadType is a meta-function for generating the
ViewFad type from a given Fad type.  When the Fad type is const, I was
putting const in the scalar type of this last template parameter, which
generated a number of build errors when compiling with the nvcc_wrapper.